# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: StockFlow — stock management, storefront, social console
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products with filters
      parameters:
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: categoryId
          schema:
            type: integer
        - in: query
          name: visible
          schema:
            type: boolean
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProductWithDetails"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductWithDetails"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductWithDetails"
    patch:
      operationId: updateProduct
      tags: [products]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductWithDetails"
    delete:
      operationId: deleteProduct
      tags: [products]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /products/{id}/variants:
    post:
      operationId: createVariant
      tags: [products]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVariantInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Variant"
  /variants/{id}:
    patch:
      operationId: updateVariant
      tags: [products]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateVariantInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Variant"
    delete:
      operationId: deleteVariant
      tags: [products]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /storefront/products:
    get:
      operationId: listStorefrontProducts
      tags: [storefront]
      summary: Public product catalog
      parameters:
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: categoryId
          schema:
            type: integer
        - in: query
          name: maxPrice
          schema:
            type: number
        - in: query
          name: inStockOnly
          schema:
            type: boolean
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StorefrontProduct"
  /storefront/products/{id}:
    get:
      operationId: getStorefrontProduct
      tags: [storefront]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StorefrontProductDetail"
...
  /storefront/contact:
    post:
      operationId: submitContactMessage
      tags: [storefront]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactMessageInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactMessage"
components:
  schemas:
...
    Variant:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        sku:
          type: string
        barcode:
          type: ["string", "null"]
        size:
          type: ["string", "null"]
        color:
          type: ["string", "null"]
        threshold:
          type: integer
        onHand:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, productId, sku, barcode, size, color, threshold, onHand, createdAt]
    CreateVariantInput:
      type: object
      properties:
        sku:
          type: string
        barcode:
          type: ["string", "null"]
        size:
          type: ["string", "null"]
        color:
          type: ["string", "null"]
        threshold:
          type: integer
        initialQuantity:
          type: integer
        locationId:
          type: ["integer", "null"]
      required: [sku, threshold]
    UpdateVariantInput:
      type: object
      properties:
        sku:
          type: string
        barcode:
          type: ["string", "null"]
        size:
          type: ["string", "null"]
        color:
          type: ["string", "null"]
        threshold:
          type: integer
    ProductWithDetails:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        reference:
          type: string
        description:
          type: ["string", "null"]
        priceHt:
          type: number
        vatRate:
          type: number
        priceTtc:
          type: number
        imageUrl:
          type: ["string", "null"]
        gallery:
          type: array
          items:
            type: string
        categoryId:
          type: ["integer", "null"]
        categoryName:
          type: ["string", "null"]
        supplierId:
          type: ["integer", "null"]
        supplierName:
          type: ["string", "null"]
        visibleOnStorefront:
          type: boolean
        updatedBy:
          type: ["string", "null"]
        updatedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        variants:
          type: array
          items:
            $ref: "#/components/schemas/Variant"
        totalOnHand:
          type: integer
      required:
        - id
        - name
        - reference
        - description
        - priceHt
        - vatRate
        - priceTtc
        - imageUrl
        - gallery
        - categoryId
        - categoryName
        - supplierId
        - supplierName
        - visibleOnStorefront
        - updatedBy
        - updatedAt
        - createdAt
        - variants
        - totalOnHand
    CreateProductInput:
      type: object
      properties:
        name:
          type: string
        reference:
          type: string
        description:
          type: ["string", "null"]
        priceHt:
          type: number
        vatRate:
          type: number
        imageUrl:
          type: ["string", "null"]
        gallery:
          type: array
          items:
            type: string
        categoryId:
          type: ["integer", "null"]
        supplierId:
          type: ["integer", "null"]
        visibleOnStorefront:
          type: boolean
        firstVariant:
          $ref: "#/components/schemas/CreateVariantInput"
      required: [name, reference, priceHt, vatRate]
    UpdateProductInput:
      type: object
      properties:
        name:
          type: string
        reference:
          type: string
        description:
          type: ["string", "null"]
        priceHt:
          type: number
        vatRate:
          type: number
        imageUrl:
          type: ["string", "null"]
        gallery:
          type: array
          items:
            type: string
        categoryId:
          type: ["integer", "null"]
        supplierId:
          type: ["integer", "null"]
        visibleOnStorefront:
          type: boolean
...
    StorefrontProduct:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        reference:
          type: string
        description:
          type: ["string", "null"]
        priceTtc:
          type: number
        imageUrl:
          type: ["string", "null"]
        categoryName:
          type: ["string", "null"]
        inStock:
          type: boolean
        totalOnHand:
          type: integer
      required: [id, name, reference, description, priceTtc, imageUrl, categoryName, inStock, totalOnHand]
    StorefrontProductDetail:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        reference:
          type: string
        description:
          type: ["string", "null"]
        priceHt:
          type: number
        priceTtc:
          type: number
        vatRate:
          type: number
        imageUrl:
          type: ["string", "null"]
        gallery:
          type: array
          items:
            type: string
        categoryName:
          type: ["string", "null"]
        inStock:
          type: boolean
        totalOnHand:
          type: integer
        variants:
          type: array
          items:
            $ref: "#/components/schemas/Variant"
      required:
        - id
        - name
        - reference
        - description
        - priceHt
        - priceTtc
        - vatRate
        - imageUrl
        - gallery
        - categoryName
        - inStock
        - totalOnHand
        - variants
    ContactMessageInput:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        message:
          type: string
      required: [name, email, message]
    ContactMessage:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        message:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, name, email, message, createdAt]